Fix for lint errors
