+++ test/e2enew/data/cni/calico/calico.yaml
