+++ test/e2enew/data/cni/kindnet/kindnet.yaml
