+++ test/e2enew/data/infrastructure-vsphere/cluster-template.yaml
