+++ test/e2enew/e2e_suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2e_new
+package e2enew
 
 import (
 	"context"
@@ -77,7 +77,6 @@ func init() {
 }
 
 func TestE2E(t *testing.T) {
-
 	// If running in prow, make sure to use the artifacts folder that will be reported in test grid (ignoring the value provided by flag).
 	if prowArtifactFolder, exists := os.LookupEnv("ARTIFACTS"); exists {
 		artifactFolder = prowArtifactFolder
